Jonathan

+++ car-rental/src/main/java/com/revature/carrental/controller/RentalsController.java
@@ -1,13 +1,13 @@
 package com.revature.carrental.controller;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
-
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,7 +19,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import com.revature.carrental.dao.RentalsDAO;
 import com.revature.carrental.model.Rentals;
-import com.revature.carrental.model.Users;
 
 
 @RestController
@@ -34,16 +33,41 @@ public class RentalsController {
 	 */
 	
 	@PostMapping("/add")
-	public Rentals createrentals(@Valid @RequestBody Rentals rentals) {
+	public ResponseEntity<Rentals> createrentals(@Valid @RequestBody Rentals rentals) {
 		rentals.setTransactionid(rentals.getTransactionid());
 		rentals.setUserid(rentals.getUserid());
 		rentals.setCarid(rentals.getCarid());
+		// make model of all current rentals for that car
+		ArrayList<Rentals> y = rentalsDAO.findAll(rentals.getCarid());
+
+		// if start is before an end AND after the start of another rentals. return invalid
+		Date inQStart = rentals.getDaterented();
+		Date inQEnd = rentals.getExpectedreturn();
+		Date listEnd;
+		Date listStart;
+		for (int x = 0; x < y.size(); x++) {
+			listEnd = y.get(x).getExpectedreturn();
+			listStart = y.get(x).getDaterented();
+			// we now have start and end of current list, and start/end of dates of request
+			if( (inQStart.compareTo(listStart) < 0 && inQStart.compareTo(listEnd) <= 0) 
+					|| 
+				(inQStart.compareTo(listStart) > 0 && inQStart.compareTo(listEnd) >= 0)) {
+					// start is before other starts, or after || = other ends
+				} else { return ResponseEntity.badRequest().build(); }
+				if( (inQEnd.compareTo(listStart) <= 0 && inQEnd.compareTo(listEnd) < 0) 
+					|| 
+				(inQEnd.compareTo(listStart) >= 0 && inQEnd.compareTo(listEnd) > 0)) {
+				// end is before || = other starts, or after other ends
+			} else { return ResponseEntity.badRequest().build(); }}
+		//if it didn't end early, rental dates are valid and may be accepted
 		rentals.setDaterented(rentals.getDaterented());
 		rentals.setExpectedreturn(rentals.getExpectedreturn());
+		//no errors from the for loop, dates may be set
 		rentals.setDescription(rentals.getDescription());
 		rentals.setApproved(rentals.getApproved());
 		
-		return rentalsDAO.save(rentals);
+		rentalsDAO.save(rentals);
+		return ResponseEntity.ok().body(rentals);
 	}
 	
 	
@@ -62,11 +86,11 @@ public class RentalsController {
 	 */
 	
 	@GetMapping("/users/query")
-	public List<Rentals> getUsersByQuery(@RequestBody Rentals rentals, Integer x){
+	public List<Rentals> getUsersByQuery(@RequestBody Rentals rentals, Integer x, Integer y, Integer z){
 		if (x == null) {
 			x = 0;
 		}
-		return rentalsDAO.findByCriteria(rentals, new PageRequest(x,10));
+		return rentalsDAO.findByCriteria(rentals, new PageRequest(x,10), y, z);
 	}
 	
 	/*
@@ -121,14 +145,16 @@ public class RentalsController {
 	public ResponseEntity<Rentals> deleterentals(@PathVariable(value="id") Long rentalid){
 		
 		Rentals rentals = rentalsDAO.findOne(rentalid);
-		
 		if(rentals == null) {
 			return ResponseEntity.notFound().build();
 		}
-		
-		rentalsDAO.delete(rentals);
-		
-		return ResponseEntity.ok().build();
-		
+		Date end;
+			end = rentals.getExpectedreturn();
+			Date now = new java.util.Date();
+			if(now.compareTo(end) > 0) {
+				return ResponseEntity.badRequest().build();
+			}
+			rentalsDAO.delete(rentals);
+			return ResponseEntity.ok().build();
 	}
 }
